requirements.txt
PyYAML==6.0
numpy==1.19.5
h5py==3.1.0
opencv-python==4.6.0.66
imageio==2.15.0
scikit-video==1.1.11
torch==1.7.1+cu110
torchvision==0.8.2+cu110
tqdm==4.64.0
Pillow==8.4.0